+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##[Unreleased]
+### Added
+- 'ofdft_ml.quantum.utils': add a new class *periodic_potential*
+
+### Changed
+
+### Removed
+- 'ofdft_ml.quantum.utils': remove the *kpotential_gen* function
+
+
 ## 0.7.0 - 2019-04-24
 ### Added
 - 'ofdft_ml.model': high level API for ML model